Extracts summaries of alignment simulations from Mplus output file

Extracts summaries of alignment simulations from Mplus output file

Usage

extractAlignmentSim(
  sim.outputs = c("sim500.out", "sim100.out", "sim1000.out"),
  silent = FALSE,
  manual = F,
  adjust = T
)

Arguments

sim.outputs

a character vector of file names containing simulation results of alignment ran in Mplus.

silent

Logical. Used for debugging.

manual

logical. If TRUE, correlations between population and estimated values are computed using the saved values (MONTECARLO:RESULTS = filename.txt should be in the Mplus input), if FALSE, the correlations are extracted from the Mplus output.

adjust

logical. Whether correlations should be adjusted using Fisher transformation (more accurate) or not (default in Mplus). Only makes sense when manual = TRUE.

Value

Invisibly returns a summary table.

Details

Best used as part of runAlignment. To match the Mplus output, the correlations are not adjusted for non-normality when averaging; when manual and adjusted, correlations are first Fisher-transformed, then averaged, then transformed back to r metric.

See also

Examples

if (FALSE) { # \dontrun{  
     align.sim.summ <- extractAlignmentSim (c("sim500.out", "sim100.out", "sim1000.out")) 
     } # }