Conventional test of measurement invariance in three steps

Computes a set of CFA models (e.g., configural, metric, and scalar/threshold invariance) and compares them to each other.

Usage

globalMI(
  ...,
  chi.sq = FALSE,
  omit = "",
  what = c("cfi", "tli", "rmsea", "srmr", "chisq"),
  partial = NULL,
  more = NULL,
  ho = F,
  quiet = FALSE
)

Arguments

...

Formula, group, and all the other eligible arguments of cfa function.

chi.sq

Logical, if chi-square should be included in the model comparison table.

omit

Character (vector). Which model should be omitted. Possible values "configural", "metric", "scalar".

what

Character (vector) of fit measures to report (see fitMeasures

partial

List of parameters to release constraints, see `group.partial` lavOptions

more

More models with extra constraints, currently limited to `residuals` and `means` which are added on the top of scalar model (or the most constrained otherwise, if 'scalar' is omitted)

ho

Currently defunct, for the higher-order factor models invariance see globalMIH

quiet

Logical, if the results should be printed.

Value

A list with all the fitted models, an LRT test lavTestLRT, and a table of fit indices (`what`) and their differences across fitted models.

Details

If the model has ordered indicators, then instead of configural-metric-scalar it fits configural-threshold-scalar sequence of models.