Test for measurement invariance for Higher-Order Factor models

Computes configural, multiple metric, and multiple scalar invariance models and compares them.

Usage

globalMIH(
  model = model,
  ...,
  chi.sq = FALSE,
  omit = "",
  what = c("cfi", "tli", "rmsea", "srmr", "chisq"),
  partial = NULL,
  more = NULL
)

Arguments

model

lavaan syntax model definition. Keep it simple as it will be appended by the function. Please use one of the loadings to identify the model at each factor order.

...

Data, group, and all the other eligible arguments of cfa function.

chi.sq

Logical, if chi-square should be reported.

omit

Currently defunct. Character (vector). Which model should be omitted. Possible values "configural", "metric", "scalar".

what

Character (vector) of fit measures to report (see fitMeasures

partial

List of parameters to release constraints, see `group.partial` of cfa

Details

For convenience, does not follow Rudnev et al. (2018) but instead the following series of constraints:

Configural All loadings are free

Metric-1 Loadings for the first-order factors are constrained to equality, first-order intercepts are free, the other means/intercepts are fixed to 0.

Metric-2, 3 etc. Applies the factor loading equality constraints to second-, third-, etc order factors.

Scalar-1 Constrains item intercepts to equality, while freely estimating the first-order factor means/intercepts.

Scalar-2, 3, etc. Sets the means/intercepts of first-order, second-order, etc. factors free while fixing to 0 all the other means/intercepts (except for the lowest level item intercepts)