Runs exact invariance tests in Mplus

Runs exact invariance tests in Mplus

Usage

runMImplus(
  model,
  group,
  data,
  categorical = NULL,
  filename = NULL,
  Mplus_com = "mplus"
)

Arguments

model

Character. Formula in Mplus format, e.g. "Factor1 BY item1 item2 item3 item4; item4 WITH item3;", see example.

group

Character, name of the grouping variable.

data

Data frame containing data.

categorical

Character vector of variable names. Indicators that are binary or ordinal. Default is NULL.

filename

Optional custom filename

Mplus_com

Sometimes you don't have a direct access to Mplus, so this argument specifies what to send to a system command line. Default value is "mplus".

Value

A summary table with fit indices and their differences between models.

Examples

if (FALSE) { # \dontrun{
   runMImplus(model = 
     "F1 BY F114 F115 F116 F117;
      F2 BY F118 F119  F120  F121 F122  F123;",
      group = "S003",
      data = dd)
} # }